Add kamal proxy update
For easy updates from Traefik to kamal-proxy, add `kamal proxy update`.

This stops and removes Traefik and kamal-proxy containers (just in case
it is run after an update). Then it starts kamal-proxy and calls
`kamal-proxy deploy` to route to any app containers that are already
running.

It can be run with a rolling option and calls the `pre-proxy-reboot` and
`post-proxy-reboot` hooks for each host.