Switch to the VFS storage driver in CI
Docker 29 broke the integration tests on GitHub Actions. It seems the
switch to the containerd image store meant we were trying to mount
one overlayfs on top of another which Linux does not currently support.

Fixed by moving to the VFS storage driver. This requires more storage so
we also need to switch the docker data root to run from /mnt/docker.