Use VFS storage driver locally
Changes in Docker 29 mean that we can't use the overlayfs storage driver
inside Docker as the outer container already uses overlayfs and Linux
doesn't support nesting.

Instead we'll use the VFS storage driver for local testing. We don't
do this on GitHub Actions - there's we've instead got the outer Docker
using vfs already and that seem to run the tests more quickly.