Boot in parallel for roles on the same host
If you have multiple roles on the same host, we'll boot them
sequentially on that host. This may be important to avoid overloading
the server but it's actually just there as a consequence of SSHKit's
`on()` method only connecting once per host.

To allow roles to boot in parallel on the same host, add a new
option:
```yaml
boot
  parallel_roles: true
```

This is implenented by adding a new `on_roles` method that uses a
separate thread for each role.

It defaults to `false` though as underpowered servers may struggle
with booting new containers in parallel.